Rename freeze.py to param_freeze.py and optuna.py to param_freeze.py to avoid name conflicts with libs; Add a default value (None) for the method parameter y_data in the GNNLearn class to fix the signature mismatch warning and simplify the class
